% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{count_authors}
\alias{count_authors}
\title{Count the number of topics by author, by theme and by date}
\usage{
count_authors(
  topics,
  by_theme = TRUE,
  by_date = c("month_year", "year", "all"),
  .sort = TRUE
)
}
\arguments{
\item{topics}{A tibble of topics by theme from en-net forum produced through
a call to \code{get_themes_topics}.}

\item{by_theme}{Logical. If TRUE (default), count by theme.}

\item{by_date}{Should topics be counted by month of the year or just by
year or total? Default is to count by month of the year.}

\item{.sort}{Logical. Should output be sorted by count frequencies? Default
is TRUE.}
}
\value{
A tibble of topic counts by author, by theme and by specified date
format
}
\description{
Count the number of topics by author, by theme and by date
}
\examples{
library(magrittr)
x <- ennet_topics
x \%>\% count_authors()

}
