% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{count_topics_day}
\alias{count_topics_day}
\alias{count_topics_week}
\alias{count_topics_month}
\alias{count_topics_year}
\title{Count number of questions/topics posted on en-net by time}
\usage{
count_topics_day(topics = get_themes_topics(), .sort = FALSE)

count_topics_week(topics = get_themes_topics(), .sort = FALSE)

count_topics_month(topics = get_themes_topics(), .sort = FALSE)

count_topics_year(topics = get_themes_topics(), .sort = FALSE)
}
\arguments{
\item{topics}{A tibble of topics by theme, by author, and by posting date
from \strong{en-net} forum produced through a call to \code{\link[=get_themes_topics]{get_themes_topics()}}}

\item{.sort}{Logical. Should output be sorted by count frequencies? Default
is \emph{FALSE}}
}
\value{
A tibble of topic counts by specified time grouping
}
\description{
Count number of questions/topics posted on en-net by time
}
\examples{
## Get counts of topics by day
count_topics_day(topics = ennet_topics)

## Get counts of topics by week
count_topics_week(topics = ennet_topics)

## Get counts of topics by month
count_topics_month(topics = ennet_topics)

## Get counts of topics by year
count_topics_year(topics = ennet_topics)

}
\author{
Ernest Guevarra
}
