% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db_topics_hourlies}
\alias{create_db_topics_hourlies}
\title{Create hourly topics datasets for the ennet_db}
\usage{
create_db_topics_hourlies(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date()
)
}
\arguments{
\item{repo}{A character value of the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}

\item{branch}{A character value for the branch name from which to retrieve
data. Default is \code{main}.}

\item{.date}{A character value or vector of date/dates for which to create
a topics dataset for the ennet_db}
}
\value{
A tibble of specified topics dataset created from data in the
ennet_db
}
\description{
Create hourly topics datasets for the ennet_db
}
\examples{
create_db_topics_hourlies(.date = "2020-12-31")

}
\author{
Ernest Guevarra
}
