% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db_topics_interactions}
\alias{create_db_topics_interactions}
\title{Create various topics interactions datasets for the ennet_db}
\usage{
create_db_topics_interactions(
  dailies,
  id = c("daily", "weekly", "monthly", "yearly")
)
}
\arguments{
\item{dailies}{A tibble of topics data usually produced by using the
\code{\link[=create_db_topics_dailies]{create_db_topics_dailies()}} function}

\item{id}{A character value for data identifier. Possible choices are
\emph{daily}, \emph{weekly}, \emph{monthly}, or \emph{yearly}.}
}
\value{
A tibble of specified topics dataset created from data in the
ennet_db
}
\description{
Create various topics interactions datasets for the ennet_db
}
\examples{
themes <- ennet_themes$themes
x <- ennet_dailies[ennet_dailies$Theme == themes[3], ]
create_db_topics_interactions(dailies = x, id = "yearly")

}
\author{
Ernest Guevarra
}
