% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db_topics_monthly}
\alias{create_db_topics_monthly}
\title{Create monthly topics dataset for the ennet_db}
\usage{
create_db_topics_monthly(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date() - 1
)
}
\arguments{
\item{repo}{A character value of the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}

\item{branch}{A character value for the branch name from which to retrieve
data. Default is \code{main}.}

\item{.date}{A character value or vector of date/dates for which to create
a topics dataset for the ennet_db}
}
\value{
A tibble of monthly topics dataset created from data in the
ennet_db
}
\description{
Create monthly topics dataset for the ennet_db
}
\examples{
create_db_topics_monthly(.date = "2021-01-01")

}
\author{
Ernest Guevarra
}
