% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennet.R
\name{ennet-deprecated}
\alias{ennet-deprecated}
\title{Deprecated functions in ennet}
\description{
These functions still work but will be removed (defunct) in the next version
of \code{ennet}.
}
\details{
\tabular{ll}{
   \strong{Function} \tab \strong{Notes} \cr
   \code{\link[=count_topics]{count_topics()}} \tab Please use \code{\link[=count_topics_theme]{count_topics_theme()}} instead \cr
   \code{\link[=count_authors]{count_authors()}} \tab Please use \code{\link[=count_topics_author]{count_topics_author()}} instead \cr
}
}
