% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ennet_dailies}
\alias{ennet_dailies}
\title{Daily extracts of topics dataset from en-net online forum}
\format{
A tibble with 90 rows and 7 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \strong{Variable} \tab \strong{Description} \cr
   :--- \tab :--- \cr
   \code{Theme} \tab Thematic areas in the en-net forum \cr
   \code{Topic} \tab Short description of the topic/question being discussed/raised \cr
   \code{Author} \tab Name of person who raised the topic/question \cr
   \code{Posted} \tab Date topic/question was posted on en-net forum \cr
   \code{Link} \tab URL of the topic/question being discussed/raised \cr
   \code{Interaction} \tab Type of interaction. Either \emph{Views} or \emph{Replies} \cr
   \code{Extraction} \tab Date and time when data was extracted \cr
   \code{n} \tab Number or count \cr
   \verb{Extraction Date} \tab Date when data was extracted \cr
}
}
\source{
https://www.en-net.org
}
\usage{
ennet_dailies
}
\description{
Daily extracts of topics dataset from en-net online forum
}
\examples{
ennet_dailies

}
\keyword{datasets}
