% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ennet_dataset}
\alias{ennet_dataset}
\title{ennet dataset}
\format{
A tibble with 8650 rows and 10 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{theme} \tab Thematic area \cr
   \code{topic} \tab Title of the question/topic \cr
   \code{user} \tab Name of ennet user either asking the question/topic or providing responses \cr
   \code{userCode} \tab Code for ennet user asking the question/topic or providing responses \cr
   \code{job} \tab Job title of ennet user asking the question/topic or providing responses \cr
   \code{role} \tab Assigned role of ennet user \cr
   \code{date_time} \tab Date and time a question/topic is raised or answer is provided \cr
   \code{type} \tab Whether a \emph{question} or an \emph{answer} \cr
   \code{code} \tab Code for the question/topic or answer \cr
   \code{post} \tab The question/topic raised or the answer provided \cr
}
}
\source{
https://www.en-net.org
}
\usage{
ennet_dataset
}
\description{
ennet dataset retrieved as of 4 October 2020
}
\examples{
ennet_dataset

}
\keyword{datasets}
