% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ennet_themes}
\alias{ennet_themes}
\title{en-net themes}
\format{
A tibble with 18 rows and 2 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{themes} \tab Thematic areas in the en-net forum \cr
   \code{links} \tab URL of the thematic area \cr
}
}
\source{
https://www.en-net.org
}
\usage{
ennet_themes
}
\description{
Themes from en-net forum retrieved on 11 October 2020
}
\examples{
ennet_themes

}
\keyword{datasets}
