% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ennet_topics}
\alias{ennet_topics}
\title{en-net topics}
\format{
A tibble with 2980 rows and 7 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{Theme} \tab Thematic areas in the en-net forum \cr
   \code{Topic} \tab Short description of the topic/question being discussed/raised \cr
   \code{Views} \tab Number of views of the topic/question being discussed/raised \cr
   \code{Author} \tab Name of person who raised the topic/question \cr
   \code{Posted} \tab Date topic/question was posted on en-net forum \cr
   \code{Link} \tab URL of the topic/question being discussed/raised \cr
   \code{Replies} \tab Number of replies to the topic/question being discussed/raised \cr
}
}
\source{
https://www.en-net.org
}
\usage{
ennet_topics
}
\description{
Topics from en-net forum retrieved on 11 October 2020
}
\examples{
ennet_topics

}
\keyword{datasets}
