% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_discussions}
\alias{get_db_discussions}
\alias{get_db_topics_daily_interactions}
\alias{get_db_topics_weekly_interactions}
\alias{get_db_topics_monthly_interactions}
\alias{get_db_topics_yearly_interactions}
\title{Retrieve data from ennet_db GitHub repository}
\usage{
get_db_discussions(repo = "katilingban/ennet_db", branch = "main")

get_db_topics_daily_interactions(
  repo = "katilingban/ennet_db",
  branch = "main"
)

get_db_topics_weekly_interactions(
  repo = "katilingban/ennet_db",
  branch = "main"
)

get_db_topics_monthly_interactions(
  repo = "katilingban/ennet_db",
  branch = "main"
)

get_db_topics_yearly_interactions(
  repo = "katilingban/ennet_db",
  branch = "main"
)
}
\arguments{
\item{repo}{A character value for the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}

\item{branch}{A character value for the branch name from which to retrieve
data. Default is \code{main}.}
}
\value{
A tibble of the specified dataset
}
\description{
Retrieve data from ennet_db GitHub repository
}
\examples{
## Retrieve discussions dataset
get_db_discussions()

## Retrieve en-net topics daily interactions dataset
get_db_topics_daily_interactions()

## Retrieve en-net topics weekly interactions dataset
get_db_topics_weekly_interactions()

## Retrieve en-net topics monthly interactions dataset
get_db_topics_monthly_interactions()

## Retrieve en-net topics yearly interactions dataset
get_db_topics_yearly_interactions()

}
\author{
Ernest Guevarra
}
