% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_themes.R
\name{get_theme_topics}
\alias{get_theme_topics}
\title{Get theme topics}
\usage{
get_theme_topics(link)
}
\arguments{
\item{link}{URL of a specific thematic area}
}
\value{
A tibble of all topics for the specified thematic area.
}
\description{
Get theme topics
}
\examples{
themes <- get_themes()
get_theme_topics(link = themes$links[1])

}
