% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_themes.R
\name{get_themes}
\alias{get_themes}
\title{Get list of thematic areas from en-net.org}
\usage{
get_themes(base = "https://www.en-net.org")
}
\arguments{
\item{base}{Base URL of the en-net site. Set to \url{https://www.en-net.org/}}
}
\value{
A tibble containing the thematic areas from en-net.org and the
  corresponding URLs for each thematic area
}
\description{
Get list of thematic areas from en-net.org
}
\examples{
get_themes()

}
