% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{get_topic_question}
\alias{get_topic_question}
\title{Get the question for a particular topic}
\usage{
get_topic_question(link)
}
\arguments{
\item{link}{URL for topic discussion}
}
\value{
A tibble containing the topic question
}
\description{
Get the question for a particular topic
}
\examples{
links <- get_theme_topics(link = get_themes()$links[4])
get_topic_question(link = links$Link)



}
