% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discussions.R
\name{get_topics_discussions}
\alias{get_topics_discussions}
\title{Get the discussion and details of discussion for a set of topics}
\usage{
get_topics_discussions(links)
}
\arguments{
\item{links}{A tibble of topics containing URL of topic discussion. This
is provided using a call to \code{get_theme_topics} or
\code{get_themes_topics}}
}
\value{
A tibble containing the topic discussions for selected topic/s
}
\description{
Get the discussion and details of discussion for a set of topics
}
\examples{
links <- get_theme_topics(link = get_themes()$links[4])
get_topics_discussions(link = links[1:3])

}
