% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ennet.R
\name{retrieve_ennet}
\alias{retrieve_ennet}
\title{Retrieve full en-net dataset}
\usage{
retrieve_ennet()
}
\value{
A tibble containing the topic discussions for all topics across all
thematic areas in en-net forum up to current point in time
}
\description{
Retrieve full en-net dataset
}
\examples{
if(interactive()) retrieve_ennet()

}
