% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ennet.R
\name{update_ennet}
\alias{update_ennet}
\title{Update full en-net dataset}
\usage{
update_ennet(freq = 0)
}
\arguments{
\item{freq}{A numeric value for time in seconds for frequency of retrieval
of updated en-net dataset. Defaults to 0 for a single retrieval
(no repeats)}
}
\value{
A tibble containing the topic discussions for all topics across all
thematic areas in en-net forum
}
\description{
Update full en-net dataset
}
\examples{
if(interactive()) update_ennet()

}
