% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_topics.R
\name{update_topics}
\alias{update_topics}
\title{Update en-net topics}
\usage{
update_topics(freq = 0)
}
\arguments{
\item{freq}{A numeric value for time in seconds for frequency of retrieval
of updated en-net topics. Defaults to 0 for a single retrieval
(no repeats)}
}
\value{
A tibble containing the topics across all thematic areas in
en-net forum
}
\description{
Update en-net topics
}
\examples{
if(interactive()) update_topics()

}
