% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{count_topics}
\alias{count_topics}
\title{Count the number of topics by theme and by date}
\usage{
count_topics(topics, by_date = c("month_year", "year"))
}
\arguments{
\item{topics}{A tibble of topics by theme from en-net forum produced through
a call to \code{get_themes_topics}.}

\item{by_date}{Should topics be counted by month of the year or just by
year?}
}
\value{
A tibble of topic counts by theme and by specified date format
}
\description{
Count the number of topics by theme and by date
}
\examples{
library(magrittr)
x <- get_themes()[4, ] \%>\% get_themes_topics()
x \%>\% count_topics(by_date = "month_year")

}
