% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db_topics_daily}
\alias{create_db_topics_daily}
\alias{create_db_topics_monthly}
\alias{create_db_topics_hourlies}
\alias{create_db_topics_dailies}
\alias{create_db_topics_daily_interactions}
\alias{create_db_topics_weekly_interactions}
\alias{create_db_topics_monthly_interactions}
\alias{create_db_topics_yearly_interactions}
\title{Create various topics dataset for the ennet_db}
\usage{
create_db_topics_daily(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date() - 1,
  fn = NULL
)

create_db_topics_monthly(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date() - 1
)

create_db_topics_hourlies(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date()
)

create_db_topics_dailies(hourlies)

create_db_topics_daily_interactions(dailies)

create_db_topics_weekly_interactions(dailies)

create_db_topics_monthly_interactions(dailies)

create_db_topics_yearly_interactions(dailies)
}
\arguments{
\item{repo}{A character value of the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}

\item{branch}{A character value for the branch name from which to retrieve
data. Default is \code{main}.}

\item{.date}{A character value or vector of date/dates for which to create
a topics dataset for the ennet_db}

\item{fn}{A character value or vector of filenames for hourly topics dataset
found in ennet_db}

\item{hourlies}{A tibble of topics data usually produced by using the
\code{\link[=create_db_topics_hourlies]{create_db_topics_hourlies()}} function}

\item{dailies}{A tibble of topics data usually produced by using the
\code{\link[=create_db_topics_dailies]{create_db_topics_dailies()}} function}
}
\value{
A tibble of specified topics dataset created from data in the
ennet_db
}
\description{
Create various topics dataset for the ennet_db
}
\examples{
##
fn <- c("ennet_topics_2021-01-17_00:54:48.csv",
        "ennet_topics_2021-01-17_02:48:38.csv",
        "ennet_topics_2021-01-17_03:57:46.csv")

create_db_topics_daily(.date = "2021-01-17", fn = fn)

create_db_topics_monthly()

create_db_topics_hourlies(.date = "2020-12-31")

create_db_topics_dailies(hourlies = ennet_hourlies)

create_db_topics_daily_interactions(dailies = ennet_dailies)

create_db_topics_weekly_interactions(dailies = ennet_dailies)

create_db_topics_monthly_interactions(dailies = ennet_dailies)

create_db_topics_yearly_interactions(dailies = ennet_dailies)

}
\author{
Ernest Guevarra
}
