% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db_topics_daily}
\alias{create_db_topics_daily}
\alias{create_db_topics_monthly}
\title{Create a daily topics dataset for the ennet_db}
\usage{
create_db_topics_daily(
  gh = "katilingban/ennet_db",
  .date = Sys.Date() - 1,
  fn = NULL
)

create_db_topics_monthly(gh = "katilingban/ennet_db", .date = Sys.Date() - 1)
}
\arguments{
\item{gh}{A character value of the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}

\item{.date}{A character value or vector of date/dates for which to create
a topics dataset for the ennet_db}

\item{fn}{A character value or vector of filenames for hourly topics dataset
found in ennet_db}
}
\description{
Create a daily topics dataset for the ennet_db
}
\examples{
##
fn <- c("ennet_topics_2021-01-17_00:54:48.csv",
        "ennet_topics_2021-01-17_02:48:38.csv",
        "ennet_topics_2021-01-17_03:57:46.csv")

create_db_topics_daily(.date = Sys.Date(), fn = fn)

create_db_topics_monthly()

}
\author{
Ernest Guevarra
}
