% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennet.R
\docType{package}
\name{ennet}
\alias{ennet}
\alias{ennet-package}
\title{Utilities to Extract Text Data From EN-NET}
\description{
en-net is the go to online forum for field practitioners requiring prompt
technical advice for operational challenges for which answers are not readily
accessible in current guidelines. The questions and the corresponding answers
raised within en-net can provide insight into what the key topics of
discussion are within the nutrition sector. This package provides utility
functions for the extraction, processing and analysis of text data from the
online forum.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/como-ph/oxcovid19}
  \item Report bugs at \url{https://github.com/como-ph/oxcovid19/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
