% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ennet_topics}
\alias{ennet_topics}
\title{Topics from en-net forum retrieved on 17 January 2021}
\format{
A tibble with 3045 rows and 7 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{Theme} \tab Thematic areas in the en-net forum \cr
   \code{Topic} \tab Short description of the topic/question being discussed/raised \cr
   \code{Views} \tab Number of views of the topic/question being discussed/raised \cr
   \code{Author} \tab Name of person who raised the topic/question \cr
   \code{Posted} \tab Date topic/question was posted on en-net forum \cr
   \code{Link} \tab URL of the topic/question being discussed/raised \cr
   \code{Replies} \tab Number of replies to topic/question being discussed/raised \cr
}
}
\source{
https://www.en-net.org
}
\usage{
ennet_topics
}
\description{
Topics from en-net forum retrieved on 17 January 2021
}
\details{
Please note that this dataset is made available in the package
primarily as a guide for the user and as testing data for the code. Users
are advised not to use this dataset for actual analysis or reporting and
instead make a call to \code{\link[=get_theme_topics]{get_theme_topics()}} or to \code{\link[=get_themes_topics]{get_themes_topics()}}.
This is because some topics and some author names have been converted to NA
as they contain non-ASCII characters which are not allowed as a text
encoding format for packaged data.
}
\examples{
ennet_topics

}
\keyword{datasets}
