% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_discussions}
\alias{get_db_discussions}
\title{Retrieve data from ennet_db GitHub repository}
\usage{
get_db_discussions(gh = "katilingban/ennet_db")
}
\arguments{
\item{gh}{A character value of the GitHub user and repository name
combination identifying the GitHub location for ennet_db. Default is
\code{katilingban/ennet_db}.}
}
\value{
A tibble of the specified dataset
}
\description{
Retrieve data from ennet_db GitHub repository
}
\examples{
## Retrieve daily topics dataset for 1 January 2021
get_db_discussions()

}
\author{
Ernest Guevarra
}
