% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_themes.R
\name{get_themes_topics}
\alias{get_themes_topics}
\title{Get theme topics}
\usage{
get_themes_topics(themes = get_themes())
}
\arguments{
\item{themes}{A tibble containing thematic areas and URLs for thematic area
pages}
}
\value{
A tibble of all topics across all thematic areas with their
  respective URLs
}
\description{
Get theme topics
}
\examples{
get_themes_topics()

}
